Define HARDENED_UNREACHABLE_() macro to invoke weakly-defined hardened_unreachable() function
Previously, the macro HARDENED_UNREACHABLE_() was hard-coded to be the always-failed
assertion 'assert(false)' in all code compiled for the host.  While this is desirable for
unit tests, there are other scenarios in which exiting gracefully or throwing a C++
exception is more appropriate.  To support these requirements, I've redefined the macro
HARDENED_UNREACHABLE_() to call a weakly-defined hardened_unreachable() function that can be
overriden with custom functionality other than the default behavior of aborting.